+++ Api.ErpFidelitas.General.v2/BusinessLogic/ParametersBL.cs
@@ -18,7 +18,7 @@ namespace Api.ErpFidelitas.General.v2.BusinessLogic
                 try
                 {
                     var Entidades = (from u in dBEntities.Parameters
-                                     where u.ParameterId == (short)id
+                                     where u.ParameterId == (int)id
                                       && u.CompanyId == CompanyId
                                      select u).FirstOrDefault();
                     if (Entidades == null)
@@ -97,7 +97,7 @@ namespace Api.ErpFidelitas.General.v2.BusinessLogic
                 try
                 {
                     var Entidades = (from u in dBEntities.Parameters
-                                     where u.ParameterId == (short)id
+                                     where u.ParameterId == (int)id
                                      && u.CompanyId == CompanyId
                                      select u).FirstOrDefault();
                     if (Entidades == null)
